requirements.txt
Flask
twilio
openai==0.28.0
python-dotenv